Code modularization - adding deequ utils
